+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/communication/chat/implementation/models/CommunicationIdentifierModel.java
@@ -9,15 +9,15 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * Identifies a participant in Azure Communication services. A participant is, for example, a phone number or an Azure
- * communication user. This model is polymorphic: Apart from kind and rawId, at most one further property may be set
- * which must match the kind enum value.
+ * communication user. This model is polymorphic: Apart from retentionPolicyKind and rawId, at most one further property may be set
+ * which must match the retentionPolicyKind enum value.
  */
 @Fluent
 public final class CommunicationIdentifierModel {
     /*
-     * The identifier kind. Only required in responses.
+     * The identifier retentionPolicyKind. Only required in responses.
      */
-    @JsonProperty(value = "kind")
+    @JsonProperty(value = "retentionPolicyKind")
     private CommunicationIdentifierModelKind kind;
 
     /*
@@ -45,18 +45,18 @@ public final class CommunicationIdentifierModel {
     private MicrosoftTeamsUserIdentifierModel microsoftTeamsUser;
 
     /**
-     * Get the kind property: The identifier kind. Only required in responses.
+     * Get the retentionPolicyKind property: The identifier retentionPolicyKind. Only required in responses.
      *
-     * @return the kind value.
+     * @return the retentionPolicyKind value.
      */
     public CommunicationIdentifierModelKind getKind() {
         return this.kind;
     }
 
     /**
-     * Set the kind property: The identifier kind. Only required in responses.
+     * Set the retentionPolicyKind property: The identifier retentionPolicyKind. Only required in responses.
      *
-     * @param kind the kind value to set.
+     * @param kind the retentionPolicyKind value to set.
      * @return the CommunicationIdentifierModel object itself.
      */
     public CommunicationIdentifierModel setKind(CommunicationIdentifierModelKind kind) {