+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/test/scala/com/azure/cosmos/spark/SparkE2EChangeFeedITest.scala
@@ -352,7 +352,7 @@ class SparkE2EChangeFeedITest
     }
 
     spark.udf.register("GetFeedRangeForPartitionKey", new GetFeedRangeForPartitionKeyValue(), StringType)
-    val pkDefinition = "{\"paths\":[\"/id\"],\"kind\":\"Hash\"}"
+    val pkDefinition = "{\"paths\":[\"/id\"],\"retentionPolicyKind\":\"Hash\"}"
     val dummyDf = spark.sql(s"SELECT GetFeedRangeForPartitionKey('$pkDefinition', '$lastId')")
 
     val feedRange = dummyDf