+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/test/scala/com/azure/cosmos/spark/SparkE2EQueryITestBase.scala
@@ -924,7 +924,7 @@ abstract class SparkE2EQueryITestBase
     df.rdd.getNumPartitions shouldEqual container.getFeedRanges.block().size()
 
     spark.udf.register("GetFeedRangeForPartitionKey", new GetFeedRangeForPartitionKeyValue(), StringType)
-    val pkDefinition = "{\"paths\":[\"/id\"],\"kind\":\"Hash\"}"
+    val pkDefinition = "{\"paths\":[\"/id\"],\"retentionPolicyKind\":\"Hash\"}"
     val dummyDf = spark.sql(s"SELECT GetFeedRangeForPartitionKey('$pkDefinition', '$lastId')")
 
     val feedRange = dummyDf