+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/communication/chat/implementation/models/ChatMessage.java
@@ -60,8 +60,8 @@ public final class ChatMessage {
     /*
      * Identifies a participant in Azure Communication services. A participant
      * is, for example, a phone number or an Azure communication user. This
-     * model is polymorphic: Apart from retentionPolicyKind and rawId, at most one further
-     * property may be set which must match the retentionPolicyKind enum value.
+     * model is polymorphic: Apart from kind and rawId, at most one further
+     * property may be set which must match the kind enum value.
      */
     @JsonProperty(value = "senderCommunicationIdentifier")
     private CommunicationIdentifierModel senderCommunicationIdentifier;
@@ -233,7 +233,7 @@ public final class ChatMessage {
     /**
      * Get the senderCommunicationIdentifier property: Identifies a participant in Azure Communication services. A
      * participant is, for example, a phone number or an Azure communication user. This model is polymorphic: Apart from
-     * retentionPolicyKind and rawId, at most one further property may be set which must match the retentionPolicyKind enum value.
+     * kind and rawId, at most one further property may be set which must match the kind enum value.
      *
      * @return the senderCommunicationIdentifier value.
      */
@@ -244,7 +244,7 @@ public final class ChatMessage {
     /**
      * Set the senderCommunicationIdentifier property: Identifies a participant in Azure Communication services. A
      * participant is, for example, a phone number or an Azure communication user. This model is polymorphic: Apart from
-     * retentionPolicyKind and rawId, at most one further property may be set which must match the retentionPolicyKind enum value.
+     * kind and rawId, at most one further property may be set which must match the kind enum value.
      *
      * @param senderCommunicationIdentifier the senderCommunicationIdentifier value to set.
      * @return the ChatMessage object itself.