+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/communication/chat/implementation/models/ChatMessageReadReceipt.java
@@ -14,8 +14,8 @@ public final class ChatMessageReadReceipt {
     /*
      * Identifies a participant in Azure Communication services. A participant
      * is, for example, a phone number or an Azure communication user. This
-     * model is polymorphic: Apart from retentionPolicyKind and rawId, at most one further
-     * property may be set which must match the retentionPolicyKind enum value.
+     * model is polymorphic: Apart from kind and rawId, at most one further
+     * property may be set which must match the kind enum value.
      */
     @JsonProperty(value = "senderCommunicationIdentifier", required = true)
     private CommunicationIdentifierModel senderCommunicationIdentifier;
@@ -37,7 +37,7 @@ public final class ChatMessageReadReceipt {
     /**
      * Get the senderCommunicationIdentifier property: Identifies a participant in Azure Communication services. A
      * participant is, for example, a phone number or an Azure communication user. This model is polymorphic: Apart from
-     * retentionPolicyKind and rawId, at most one further property may be set which must match the retentionPolicyKind enum value.
+     * kind and rawId, at most one further property may be set which must match the kind enum value.
      *
      * @return the senderCommunicationIdentifier value.
      */
@@ -48,7 +48,7 @@ public final class ChatMessageReadReceipt {
     /**
      * Set the senderCommunicationIdentifier property: Identifies a participant in Azure Communication services. A
      * participant is, for example, a phone number or an Azure communication user. This model is polymorphic: Apart from
-     * retentionPolicyKind and rawId, at most one further property may be set which must match the retentionPolicyKind enum value.
+     * kind and rawId, at most one further property may be set which must match the kind enum value.
      *
      * @param senderCommunicationIdentifier the senderCommunicationIdentifier value to set.
      * @return the ChatMessageReadReceipt object itself.