+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/main/scala/com/azure/cosmos/spark/CosmosPatchHelper.scala
@@ -36,7 +36,7 @@ class CosmosPatchHelper(diagnosticsConfig: DiagnosticsConfig,
   assertNotNullOrEmpty(path, "path")
   assertNotNull(partitionKeyDefinition, "partitionKeyDefinition")
 
-  // There are some properties are immutable, these retentionPolicyKind properties include:
+  // There are some properties are immutable, these kind properties include:
   // 1. System properties : _ts, _rid, _etag
   // 2. id, and partitionKeyPath
   if ((path.startsWith("/") && !systemProperties.contains(path.substring(1)) && IdAttributeName != path.substring(1))