+++ sdk/applicationinsights/azure-resourcemanager-applicationinsights/CHANGELOG.md
@@ -20,12 +20,12 @@
 
 #### `models.WorkbookUpdateParameters` was modified
 
-* `models.SharedTypeKind retentionPolicyKind()` -> `models.WorkbookUpdateSharedTypeKind retentionPolicyKind()`
+* `models.SharedTypeKind kind()` -> `models.WorkbookUpdateSharedTypeKind kind()`
 * `withKind(models.SharedTypeKind)` was removed
 
 #### `models.WorkbookResource` was modified
 
-* `models.Kind retentionPolicyKind()` -> `models.WorkbookSharedTypeKind retentionPolicyKind()`
+* `models.Kind kind()` -> `models.WorkbookSharedTypeKind kind()`
 * `withKind(models.Kind)` was removed
 
 #### `models.Workbook$Definition` was modified
@@ -38,7 +38,7 @@
 
 #### `models.Workbook` was modified
 
-* `models.Kind retentionPolicyKind()` -> `models.WorkbookSharedTypeKind retentionPolicyKind()`
+* `models.Kind kind()` -> `models.WorkbookSharedTypeKind kind()`
 
 ### Features Added